spaminness function

+++ Spam.hs
@@ -25,7 +25,7 @@ not_relevent = spam
 train = SpamClassificationData
 
 data SpamClassificationDict = SpamClassificationDict {
-    spamDict :: M.Map String Int,
-    hamDict :: M.Map String Int,
+    spamDict :: M.Map String Double,
+    hamDict :: M.Map String Double,
     smooth_k_ :: Double
 }
